feat(components): [el-input] refinement for code
* Extract setup to seperated composition hook files.
* Update template accordingly.
